Moved tags.js functionality into a tag-utils.js file, exported functions and added as imports to tags.js and tag-selector. Basics working
