+++ tools/caas-tag-selector/index.html
@@ -15,6 +15,8 @@
     </style>
   </head>
   <body>
-    <main></main>
+    <main>
+      <da-tag-selector></da-tag-selector>
+    </main>
   </body>
 </html>
