+++ tools/caas-tag-selector/tag-selector.css
@@ -1,0 +1,5 @@
+.tag-selector-group {
+  margin: 20px;
+  display: flex;
+  flex-direction: column;
+}
