+++ tools/caas-tag-selector/tag-selector.js
@@ -1,8 +1,34 @@
 /* eslint-disable no-underscore-dangle, import/no-unresolved */
 import { LitElement, html, nothing } from 'https://da.live/nx/deps/lit/lit-core.min.js';
 import getStyle from 'https://da.live/nx/utils/styles.js';
+import DA_SDK from 'https://da.live/nx/utils/sdk.js';
+import { getTags, getAemRepo, getRootTags } from '../tags/tag-utils.js';
 
 const style = await getStyle(import.meta.url);
+const UI_TAG_PATH = '/ui#/aem/aem/tags';
+
+const { context, actions, token } = await DA_SDK.catch(() => null);
+
+const opts = { headers: { Authorization: `Bearer ${token}` } };
+const aemConfig = await getAemRepo(context, opts).catch(() => null);
+if (!aemConfig || !aemConfig.aemRepo) {
+  console.log('error');
+  // return;
+}
+
+const namespaces = aemConfig?.namespaces.split(',').map((namespace) => namespace.trim()) || [];
+const rootTags = await getRootTags(namespaces, aemConfig, opts);
+
+if (!rootTags || rootTags.length === 0) {
+  console.log('error');
+  // return;
+}
+
+for (const tag of rootTags) {
+  console.log(tag, rootTags);
+  const tagList = await getTags(tag.path, opts);
+  console.log(tagList);
+}
 
 class DaTagSelector extends LitElement {
   static properties = {
@@ -19,7 +45,7 @@ class DaTagSelector extends LitElement {
   }
 
   render() {
-    if (this._tags.length === 0) return nothing;
+    // if (this._tags.length === 0) return nothing;
     return html`
       <div class="tag-selector-group">
         <label for"caas-content-type">CaaS Content Type</label>
